requirements.txt
click==8.1.3 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
flask-sqlalchemy==3.0.3 ; python_version >= "3.11" and python_version < "4.0"
flask==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
psycopg2==2.9.5 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.4 ; python_version >= "3.11" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.11" and python_version < "4.0"
unidecode==1.3.6 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
